GF-20 Move event classes
