+++ src/main/events/Event.java
@@ -1,4 +1,4 @@
-package main;
+package main.events;
 
 public enum Event {
     GridChanged, GridEnergized
