+++ src/main/events/IEventListener.java
@@ -1,4 +1,6 @@
-package main;
+package main.events;
+
+import main.events.Event;
 
 public interface IEventListener {
     void handleEvent(Event event);
